ref(data-scrubbing): Update tooltip description
