--- static/app/components/events/meta/annotatedText/chunk.tsx
@@ -1,24 +1,0 @@
-import Tooltip from 'sentry/components/tooltip';
-import {ChunkType} from 'sentry/types';
-
-import Redaction from './redaction';
-import {getTooltipText} from './utils';
-
-type Props = {
-  chunk: ChunkType;
-};
-
-const Chunk = ({chunk}: Props) => {
-  if (chunk.type === 'redaction') {
-    const title = getTooltipText({rule_id: chunk.rule_id, remark: chunk.remark});
-    return (
-      <Tooltip title={title}>
-        <Redaction>{chunk.text}</Redaction>
-      </Tooltip>
-    );
-  }
-
-  return <span>{chunk.text}</span>;
-};
-
-export default Chunk;
